requirements.txt
python-telegram-bot
flask
phonenumbers
